Speedup and Cleanup
Most imporantly this commit speeds up calc_solar_geom by vectorizing the inner loop.  The output has been verified to match the old loop method.  This results in ~10x speedup of the calculation.

Next, due to the inclusion of the solar_geom dictionary some of the mtclim function calls could be made cleaner.