Addressing many issues.
First, this cleans up vapor pressure and temperature disaggregation.  I've also added more disaggregation variables, such as relative humidity and cloud cover.  This mostly addresses the issues that have been present in the vapor pressure calculation, though a bit of an issue still occurs due to the calculation of .  There is still an underestimation of longwave that I haven't been able to fully track down.

Second, this speeds up the calculation of the solar geometry via numba: error: the following arguments are required: filename.  With this improvement I went from ~43 seconds for 81 grid cells to ~19 seconds on an i5 3317u.  Because of the requirements for numba to provide speedups the algorithm as been reverted to be purely imperative, and the configuration module is not arranged into multiple dictionaries.  Because some of the values in the configuration module were modified previously, they have been pulled out into the  bit of the  object.