Adjusting input longitude values to be within [-180, +180] range
Using the `math.remainder` built-in function, all the input longitude
values are transformed to be within the [-180, +180] range while
disaggregation of data is being implemented.

This quick correction is meant to correct an issue reported by Xiao Luo
<luox@hawaii.edu> with the UTC-offset option while the longitude values
are not within the [-180, +180] range, causing inaccurate results,
specifically, inaccurate time shifted values.

Reported-by: Xiao Luo <luox@hawaii.edu>
Co-authored-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>